+++ src/projects/MyNatsClient/Events/ClientWorkerFailed.cs
@@ -2,12 +2,12 @@ using System;
 
 namespace MyNatsClient.Events
 {
-    public class ClientConsumerFailed : IClientEvent
+    public class ClientWorkerFailed : IClientEvent
     {
         public INatsClient Client { get; }
         public Exception Exception { get; }
 
-        public ClientConsumerFailed(INatsClient client, Exception exception)
+        public ClientWorkerFailed(INatsClient client, Exception exception)
         {
             Client = client;
             Exception = exception;